set middlewares to http services
